backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6

# HTTP client for AI service
httpx==0.25.2

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Redis for caching (if needed)
redis==5.0.1

# Optional: Celery for background tasks (if needed)
celery==5.3.4